+++ tests/interactive/test-tar_watch.R
@@ -11,9 +11,9 @@ tar_script({
 })
 # Should just launch the app in a blocking process:
 tar_watch(background = FALSE)
-# Should exclude data1
+# Should exclude data1 in the graph.
 tar_watch(background = FALSE, exclude = "data1")
-# Should exclude sleep_run() and data2
+# Should exclude sleep_run() and data2 in the graph.
 px <- tar_watch(background = TRUE, exclude = c("sleep_run", "data2"))
 px$kill()
 # Should not block the main process: